+++ teachers/models.py
@@ -1,7 +1,5 @@
 from django.db import models
 
-from .validators import phone_validator
-
 
 class Teacher(models.Model):
 
@@ -14,7 +12,7 @@ class Teacher(models.Model):
         max_length=20,
         blank=True,
         null=True,
-        validators=[phone_validator]
+        unique=True
     )
 
     def __str__(self):
